# a bit of performance improvements