--- a/news/177.feature.1
+++ b/news/177.feature.1
@@ -1,0 +1,2 @@
+removed marking request for disable CSRF protection (need plone.scale with https://github.com/plone/plone.scale/pull/58)
+[mamico]