--- a/news/177.feature.2
+++ b/news/177.feature.2
@@ -1,0 +1,2 @@
+use the attribute _sizes for local caching correctly available_sizes
+[mamico]