--- a/news/177.feature.3
+++ b/news/177.feature.3
@@ -1,0 +1,2 @@
+add additional infos in scale storage only if missing
+[mamico]