--- a/plone/namedfile/scaling.py
+++ b/plone/namedfile/scaling.py
@@ -10,7 +10,7 @@ from plone.namedfile.interfaces import IStableImageScale
 from plone.namedfile.utils import getHighPixelDensityScales
 from plone.namedfile.utils import set_headers
 from plone.namedfile.utils import stream_data
-from plone.protect.interfaces import IDisableCSRFProtection
+from plone.protect.utils import safeWrite
 from plone.rfc822.interfaces import IPrimaryFieldInfo
 from plone.scale.interfaces import IImageScaleFactory
 from plone.scale.interfaces import IScaledImageQuality
@@ -402,7 +402,7 @@ class ImageScaling(BrowserView):
             return image.tag()
         raise TraversalError(self, name)
 
-    _sizes = {}
+    _sizes = None
 
     @deprecate("use property available_sizes instead")
     def getAvailableSizes(self, fieldname=None):
@@ -416,13 +416,14 @@ class ImageScaling(BrowserView):
     @property
     def available_sizes(self):
         # fieldname is ignored by default
-        sizes_util = queryUtility(IAvailableSizes)
-        if sizes_util is None:
-            return self._sizes
-        sizes = sizes_util()
-        if sizes is None:
-            return {}
-        return sizes
+        if self._sizes is None:
+            sizes_util = queryUtility(IAvailableSizes)
+            if sizes_util is None:
+                if self._sizes is None:
+                    self._sizes = {}
+            else:
+                self._sizes = sizes_util() or {}
+        return self._sizes
 
     @available_sizes.setter
     def available_sizes(self, value):
@@ -492,8 +493,6 @@ class ImageScaling(BrowserView):
             if scale not in available:
                 return None  # 404
             width, height = available[scale]
-        if IDisableCSRFProtection and self.request is not None:
-            alsoProvides(self.request, IDisableCSRFProtection)
         storage = getMultiAdapter(
             (self.context, functools.partial(self.modified, fieldname)),
             IImageScaleStorage
@@ -509,16 +508,20 @@ class ImageScaling(BrowserView):
         if info is None:
             return  # 404
 
-        info["srcset"] = self.calculate_srcset(
-            fieldname=fieldname,
-            height=height,
-            width=width,
-            direction=direction,
-            scale=scale,
-            storage=storage,
-            **parameters
-        )
-        info["fieldname"] = fieldname
+        if "srcset" not in info:
+            info["srcset"] = self.calculate_srcset(
+                fieldname=fieldname,
+                height=height,
+                width=width,
+                direction=direction,
+                scale=scale,
+                storage=storage,
+                **parameters
+            )
+        if "fieldname" not in info:
+            info["fieldname"] = fieldname
+        if self.request is not None:
+            safeWrite(info, self.request)
         scale_view = self._scale_view_class(self.context, self.request, **info)
         return scale_view