--- a/plone/namedfile/tests/test_scaling.py
+++ b/plone/namedfile/tests/test_scaling.py
@@ -387,10 +387,11 @@ class ImageScalingTests(unittest.TestCase):
         self.assertEqual(foo.width, 42)
         self.assertEqual(foo.height, 42)
 
-    def testAvailableSizes(self):
+    def testDefaultAvailableSizes(self):
         # by default, no named scales are configured
         self.assertEqual(self.scaling.available_sizes, {})
 
+    def testCustomAvailableSizes(self):
         # a callable can be used to look up the available sizes
         def custom_available_sizes():
             return {'bar': (10, 10)}