1. removed marking request for disable CSRF protection (scalesdict are already "safe type" for plone.protect)
2. use the attribute _sizes for local caching correctly available_sizes
3. add additional infos in scale storage only if missing (probably could be better to move writes in plone.scale)

Tested against a @querystring-search restapi that returns fullobjects with leadimage:
before 0.12289s, with these changes 0.06316s (b_size=10) and from 2.02214s to 0.97613s (b_size=200)