@mamico 2022-06-09
@mauritsvanrees @cekk 
I saw that #113 was merged. So I synced the master code in the branch of this PR and reevaluated the performance improvements.
Same, good results as above. Surprisingly, most performance improvements come from the remove of `IDisableCSRFProtection`. 
But by removing that there is a need to modify the storage adapter in `plone.scale`, just for the purpose of testing this PR, I inserted a temporary monkey patch for this (https://github.com/plone/plone.namedfile/pull/117/commits/f71706a2f9b405b4902b658a75837c405abd5401). Eventually, if it's ok in the whole, I will move the monkey patch to a related PR in `plone.scale`.