@ale-rt 2022-06-09
Just my 2 c.

I think ScalesDict is not performant because it inherits from PersistentDict:
- https://github.com/plone/plone.scale/blob/258731910a35f37056a2097d7607e35db964fad1/plone/scale/storage.py#L69

AFAIR PersistentDict are completely rewritten when something changes.

I guess that using OOBTree directly is a faster option (which comes with builtin conflict resolution that does not require you add the plone.scale to the zeo script).