@ale-rt 2022-06-09
See also the discussion happening here https://github.com/plone/plone.scale/issues/41