@mamico 2022-06-10
> Just my 2 c.
> 
> I think ScalesDict is not performant because it inherits from PersistentDict...

@alert, thank you for your suggestion. But I assume it’s another story: here the main performance improvement is not about storing/reading data, but the removal of `IDisableCSRFProtection` (which should also be a possible security improvement).