@ale-rt 2022-06-13
> > Just my 2 c.
> > I think ScalesDict is not performant because it inherits from PersistentDict...
> 
> @Alert, thank you for your suggestion. But I assume it’s another story: here the main performance improvement is not about storing/reading data, but the removal of `IDisableCSRFProtection` (which should also be a possible security improvement).

Yeah, I see your point, but I think that if performance is what you are looking for switching away from `PersistentDict` is definitely something worth considering.

Anyway, it looks good but of course this is not merge-able with that patch.