@mamico 2022-06-13
> Anyway, it looks good but of course this is not merge-able with that patch.

Absolutely, it's what I've tried to  say before. If the main idea is acceptable to all reviewers, I'll open a second PR in `plone.scale`. If you prefer, I can also remove the other optimizations except the `IDisableCSRFProtection` removing.