@mauritsvanrees 2022-06-13
I have left some comments on the commit with the temporary monkey patch (f71706a2f9b405b4902b658a75837c405abd5401).

The approach seems good. Please continue. :-)