@mauritsvanrees 2022-06-13
Thanks, sounds logical now.
I have merged your `plone.scale` PR.