@mauritsvanrees 2022-06-13
I have merged this, but GitHub UI does not show it, seems a bit slow the last few minutes.

Anyway, thanks for your work @mamico and thanks for the review @ale-rt !